move Settings.js and Settings.sc.js to the settings_pages folder
