+++ src/NoSidebarRouter.js
@@ -4,7 +4,7 @@ import ArticlesRouter from "./articles/_ArticlesRouter";
 import WordsRouter from "./words/_WordsRouter";
 import ExercisesRouter from "./exercises/ExercisesRouter";
 import TeacherRouter from "./teacher/_routing/_TeacherRouter";
-import Settings from "./pages/Settings";
+import Settings from "./pages/settings_pages/Settings";
 import ArticleReader from "./reader/ArticleReader";
 import UserDashboard from "./userDashboard/UserDashboard";
 import React from "react";
